requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Face Recognition
facenet-pytorch==2.5.3
mtcnn==0.1.1
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Image Processing
scikit-learn==1.3.2
scipy==1.11.4

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.21.1